+++ app/src/main/AndroidManifest.xml
@@ -13,11 +13,11 @@
         android:theme="@style/Theme.NOWSOPTAndroid"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".test.presentation.MainActivity"
             android:exported="false">
         </activity>
         <activity
-            android:name=".LoginActivity"
+            android:name=".test.presentation.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -26,7 +26,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".SignUpActivity"
+            android:name=".test.presentation.SignUpActivity"
             android:exported="false"/>
     </application>
 </manifest>
