+++ app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingHorizontal="30dp"
-    tools:context=".LoginActivity">
+    tools:context=".test.presentation.LoginActivity">
 
     <TextView
         android:id="@+id/tv_login_title"
