+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".test.presentation.MainActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fcv_home"
