+++ app/src/main/res/layout/activity_sign_up.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingHorizontal="30dp"
-    tools:context=".SignUpActivity">
+    tools:context=".test.presentation.SignUpActivity">
 
     <TextView
         android:id="@+id/tv_sign_up_title"
