+++ app/src/main/res/layout/fragment_mypage.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingHorizontal="30dp"
-    tools:context=".MainActivity">
+    tools:context=".test.presentation.MainActivity">
 
     <ImageView
         android:id="@+id/iv_my_profile"
