+++ app/src/main/res/layout/fragment_search.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".test.SearchFragment">
+    tools:context=".test.presentation.SearchFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_friends"
